# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Wandelbots NOVA - Isaac Sim Extension API",
        "description": "This extension enables seamless connection between NVIDIA Isaac Sim\u2122 and Wandelbots NOVA. Wandelbots NOVA simplifies the programming of industrial robots and cobots from multiple brands, allowing users to easily configure various robot models and teach them through an intuitive interface or by leveraging their preferred programming languages via APIs. Start programming your favorite robot brands like ABB, FANUC, KUKA, Universal Robots and Yaskawa in an Omniverse simulation scene, benefitting from its realistic behaviour.\n\nThe API provides comprehensive capabilities including:\n- **Stage Management**: Load, save, and manipulate USD stages\n- **Prims & Scene Graph**: Create and modify scene primitives and hierarchies\n- **Cameras**: Configure virtual cameras and viewports for visualization\n- **Motion Groups**: Define and control robot motion groups and kinematics\n- **Teaching & Trajectory**: Record, playback, and execute robot trajectories\n- **Collision Detection**: Manage collision worlds and colliders for safe path planning\n- **Nucleus Integration**: Access Omniverse Nucleus storage and collaboration features\n- **UI Control**: Interact with the Isaac Sim user interface programmatically",
        "version": "2.34.5"
    },
    "paths": {
...
        "/auth/token": {
            "post": {
                "summary": "Authenticate",
                "description": "This Endpoint allows you to authenticate with Wandelbots NOVA using your access token.",
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/Auth0Credentials"
                                    },
                                    {
                                        "$ref": "#/components/schemas/EntraIDCredentials"
                                    }
                                ],
                                "title": "Credentials",
                                "description": "Authentication credentials from NOVA"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Authenticated successfully"
                    },
                    "400": {
                        "description": "Authentication failed"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Auth0Credentials": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Provide the id to find the matching Auth0 auth configuration.",
                        "default": "default"
                    },
                    "access_token": {
                        "type": "string",
                        "title": "Access Token",
                        "description": "Provide the access token obtained for authentication"
                    }
                },
                "type": "object",
                "required": [
                    "access_token"
                ],
                "title": "Auth0Credentials"
            },
...
            "EntraIDCredentials": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Provide the id to find the matching Entra ID auth configuration."
                    },
                    "access_token": {
                        "type": "string",
                        "title": "Access Token",
                        "description": "Provide the access token obtained for authentication"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "access_token"
                ],
                "title": "EntraIDCredentials"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}